requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.16.0
numpy<2.0.0
scipy
sounddevice
soundfile
librosa
opencv-python
faster-whisper
llama-cpp-python
onnxruntime-gpu
pyaudio
transformers
accelerate
sentencepiece
huggingface_hub<0.25.0
omegaconf
einops
diffusers==0.24.0
